+++ src/CommandBus/CommandBusBundle.php
@@ -1,0 +1,38 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
+namespace Communitales\Component\CommandBus;
+
+use Communitales\Component\CommandBus\Attribute\AsCommandHandler;
+use Communitales\Component\CommandBus\DependencyInjection\CommandHandlerCompilerPass;
+use Override;
+use Symfony\Component\DependencyInjection\ChildDefinition;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\HttpKernel\Bundle\Bundle;
+
+final class CommandBusBundle extends Bundle
+{
+    #[Override]
+    public function build(ContainerBuilder $container): void
+    {
+        parent::build($container);
+
+        $container->registerAttributeForAutoconfiguration(
+            AsCommandHandler::class,
+            static function (ChildDefinition $definition, AsCommandHandler $attribute): void {
+                $definition->addTag(CommandHandlerCompilerPass::HANDLER_TAG, [
+                    'handles' => $attribute->handles,
+                ]);
+            }
+        );
+
+        $container->addCompilerPass(new CommandHandlerCompilerPass());
+    }
+}